refactor(chat): split file + channel-admin routes into chat_files.py + chat_admin.py (B5) (#556)
