+++ tinyagentos/routes/__init__.py
@@ -130,6 +130,10 @@ def register_all_routers(app):
 
     from tinyagentos.routes.chat import router as chat_router
     app.include_router(chat_router)
+    from tinyagentos.routes.chat_files import router as chat_files_router
+    app.include_router(chat_files_router)
+    from tinyagentos.routes.chat_admin import router as chat_admin_router
+    app.include_router(chat_admin_router)
 
     from tinyagentos.routes.canvas import router as canvas_router
     app.include_router(canvas_router)
